+++ logger/include/parser.h
@@ -9,12 +9,12 @@ bool filterLog(eLogLevel lvl, eLogSubsystem ss);
 // Parsing functions
 extern const ParseFunctionPointer parseFunctions[NUM_TYPES];
 // Default functions
-void parseTimeLog(uint8_t len, const char *data);
-void parseCountLog(uint8_t len, const char *data);
-void parseVersionLog(uint8_t len, const char *data);
-void parsePrintLog(uint8_t len, const char *data);
+void parseTimeLog(uint16_t len, const char *data);
+void parseCountLog(uint16_t len, const char *data);
+void parseVersionLog(uint16_t len, const char *data);
+void parsePrintLog(uint16_t len, const char *data);
 // User functions
-void parseLogStuff(uint8_t len, const char *data);
+void parseLogStuff(uint16_t len, const char *data);
 
 #define cpyRawValue(d, o) \
 	memcpy((void *)&(d), (void *)(o), sizeof(d))
